@@ -29,7 +29,7 @@
     "!dist/**/*.spec.*"
   ],
   "peerDependencies": {
-    "svelte": "^4.0.0",
+    "svelte": "^5.0.0",
     "tailwind-merge": "^1.13.2",
     "tailwindcss": "^3.3.2"
   },
@@ -38,7 +38,7 @@
     "@sveltejs/adapter-auto": "^3.3.1",
     "@sveltejs/kit": "^2.7.3",
     "@sveltejs/package": "^2.2.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/typography": "^0.5.9",
     "@typescript-eslint/eslint-plugin": "^6.2.1",
     "@typescript-eslint/parser": "^6.2.1",
@@ -47,16 +47,16 @@
     "eslint-config-prettier": "^8.9.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "postcss": "^8.4.27",
-    "prettier": "^3.0.0",
-    "prettier-plugin-svelte": "^3.0.3",
+    "prettier": "^3.1.0",
+    "prettier-plugin-svelte": "^3.2.6",
     "publint": "^0.2.0",
     "sass": "^1.64.2",
-    "svelte": "^4.1.2",
-    "svelte-check": "^3.4.6",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwind-merge": "^1.14.0",
     "tailwindcss": "^3.3.3",
     "tslib": "^2.6.1",
-    "typescript": "^5.1.6",
+    "typescript": "^5.5.0",
     "vite": "^5.4.10",
     "vitest": "^2.1.3"
   },
@@ -2375,9 +2375,9 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@floating-ui/dom": "^1.5.1",
-    "flowbite": "^1.8.1",
-    "flowbite-svelte": "^0.40.2",
-    "svelte-meta-tags": "^2.8.0"
+    "@floating-ui/dom": "^1.6.11",
+    "flowbite": "^2.5.2",
+    "flowbite-svelte": "^0.47.1",
+    "svelte-meta-tags": "^4.0.4"
   }
 }
